# Further optimise UTxO index construction.