## Issue

ADP-2975

## Summary

Applying this PR reduces the evaluation time of `transactionFee` by approximately 30% when applied to a wallet with 1,000 ada-only UTxO entries.

## Details

This PR moves the construction of all UTxO map objects:
- **_out_** of the `balanceTransaction` function;
- **_into_** a new function `constructUTxOIndex`, which is now also responsible for constructing the index.

The result of calling `constructUTxOIndex` can be fully evaluated **_before_** calling `balanceTx`, **_outside_** of any fee evaluation loop, and passed into `balanceTx`. This means that repeated calls to `balanceTx` will not result in repeated construction of UTxO maps, which are all constant in the context of the `transactionFee` evaluation loop.

Function `constructUTxOIndex` returns an object of type `Write.Tx.Balance.UTxOIndex era`, which:
- internally wraps the three types of UTxO map and index-related objects currently needed by `balanceTx`.
- is completely opaque to callers outside of the `Write.Tx.Balance` module, which means future PRs are free to change the internal definition in a way that suits `balanceTx`, without breaking the API.

## Benchmark data

All benchmark data produced from binaries built with `-O2`.

### Before applying this PR

```
    Latencies for 2 fixture wallets with 100 utxos scenario
        postTransactionFee  - 59.4 ms
        postTransactionFee  - 57.6 ms
        postTransactionFee  - 55.7 ms
        postTransactionFee  - 56.6 ms
    Latencies for 2 fixture wallets with 200 utxos scenario
        postTransactionFee  - 67.9 ms
        postTransactionFee  - 70.3 ms
        postTransactionFee  - 70.9 ms
        postTransactionFee  - 72.6 ms
    Latencies for 2 fixture wallets with 500 utxos scenario
        postTransactionFee  - 104.3 ms
        postTransactionFee  - 103.4 ms
        postTransactionFee  - 105.9 ms
        postTransactionFee  - 101.9 ms
    Latencies for 2 fixture wallets with 1000 utxos scenario
        postTransactionFee  - 197.9 ms
        postTransactionFee  - 172.4 ms
        postTransactionFee  - 189.8 ms
        postTransactionFee  - 178.1 ms
```

### After applying this PR
```
    Latencies for 2 fixture wallets with 100 utxos scenario
        postTransactionFee  - 55.2 ms
        postTransactionFee  - 62.7 ms
        postTransactionFee  - 54.4 ms
        postTransactionFee  - 57.2 ms
    Latencies for 2 fixture wallets with 200 utxos scenario
        postTransactionFee  - 57.7 ms
        postTransactionFee  - 60.5 ms
        postTransactionFee  - 62.2 ms
        postTransactionFee  - 57.9 ms
    Latencies for 2 fixture wallets with 500 utxos scenario
        postTransactionFee  - 76.8 ms
        postTransactionFee  - 92.3 ms
        postTransactionFee  - 83.0 ms
        postTransactionFee  - 79.2 ms
    Latencies for 2 fixture wallets with 1000 utxos scenario
        postTransactionFee  - 128.1 ms
        postTransactionFee  - 115.0 ms
        postTransactionFee  - 138.3 ms
        postTransactionFee  - 123.5 ms
```